+++ src/Components/LeadsComponent/LeadsKanbanFilter.jsx
@@ -1,13 +1,10 @@
-import { Tabs, Flex, Button, TextInput, Select, MultiSelect } from "@mantine/core";
-import { DatePickerInput } from "@mantine/dates";
+import { Tabs, Flex, Button } from "@mantine/core";
 import { useEffect, useState } from "react";
 import { SelectWorkflow } from "../SelectWorkflow";
 import { getLanguageByKey } from "../utils";
 import { TicketFormTabs } from "../TicketFormTabs";
 import { filteredWorkflows } from "./utils";
-import { useGetTechniciansList } from "../../hooks";
-import { MESSAGES_TYPE_OPTIONS, DD_MM_YYYY_DASH } from "../../app-constants";
-import dayjs from "dayjs";
+import { MessageFilterForm } from "./MessageFilterForm";
 
 export const LeadsKanbanFilter = ({
   onClose,
@@ -18,60 +15,10 @@ export const LeadsKanbanFilter = ({
   onSubmitTicket,
 }) => {
   const [systemWorkflow, setSystemWorkflow] = useState(filteredWorkflows);
-  const [message, setMessage] = useState("");
-  const [mtype, setMtype] = useState(null);
-  const [senderIds, setSenderIds] = useState([]);
-  const [timeSent, setTimeSent] = useState([null, null]);
-
-  const { technicians } = useGetTechniciansList();
 
   useEffect(() => {
     setSystemWorkflow(baseSystemWorkflow);
   }, [baseSystemWorkflow]);
-
-  useEffect(() => {
-    if (initialData && typeof initialData === "object") {
-      setMessage(initialData.message || "");
-      setMtype(initialData.mtype || null);
-      setSenderIds(
-        Array.isArray(initialData.sender_id)
-          ? initialData.sender_id.map(String)
-          : typeof initialData.sender_id === "string"
-            ? initialData.sender_id.split(",")
-            : []
-      );
-      if (initialData.time_sent?.from || initialData.time_sent?.to) {
-        setTimeSent([
-          initialData.time_sent?.from ? dayjs(initialData.time_sent.from, DD_MM_YYYY_DASH).toDate() : null,
-          initialData.time_sent?.to ? dayjs(initialData.time_sent.to, DD_MM_YYYY_DASH).toDate() : null,
-        ]);
-      } else {
-        setTimeSent([null, null]);
-      }
-    }
-  }, [initialData]);
-
-  const handleApplyMessageFilter = () => {
-    const filters = {};
-    if (message) filters.message = message;
-    if (mtype) filters.mtype = mtype;
-    if (senderIds.length) filters.sender_id = senderIds.map((id) => parseInt(id, 10));
-    if (timeSent?.[0] || timeSent?.[1]) {
-      filters.time_sent = {
-        ...(timeSent[0] && { from: dayjs(timeSent[0]).format(DD_MM_YYYY_DASH) }),
-        ...(timeSent[1] && { to: dayjs(timeSent[1]).format(DD_MM_YYYY_DASH) }),
-      };
-    }
-    onSubmitTicket(filters, "message");
-  };
-
-  const handleResetMessageFilter = () => {
-    setMessage("");
-    setMtype(null);
-    setSenderIds([]);
-    setTimeSent([null, null]);
-    onSubmitTicket({}, "message");
-  };
 
   return (
     <Tabs h="100%" className="leads-modal-filter-tabs" defaultValue="filter_workflow">
@@ -94,13 +41,13 @@ export const LeadsKanbanFilter = ({
           <Flex direction="column" justify="space-between" h="100%">
             <SelectWorkflow selectedValues={systemWorkflow} onChange={setSystemWorkflow} />
             <Flex justify="end" gap="md" mt="md">
-              <Button variant="outline" onClick={handleResetMessageFilter}>
+              <Button variant="outline" onClick={() => onSubmitTicket({}, "message")}>
                 {getLanguageByKey("Reset filter")}
               </Button>
               <Button variant="default" onClick={onClose}>
                 {getLanguageByKey("Închide")}
               </Button>
-              <Button variant="filled" loading={loading} onClick={handleApplyMessageFilter}>
+              <Button variant="filled" loading={loading} onClick={() => onSubmitTicket({}, "message")}>
                 {getLanguageByKey("Aplică")}
               </Button>
             </Flex>
@@ -118,55 +65,12 @@ export const LeadsKanbanFilter = ({
       </Tabs.Panel>
 
       <Tabs.Panel value="filter_message" pt="xs">
-        <Flex direction="column" gap="md">
-          <TextInput
-            label={getLanguageByKey("searchByMessages")}
-            placeholder={getLanguageByKey("searchByMessages")}
-            value={message}
-            onChange={(e) => setMessage(e.currentTarget.value)}
-          />
-
-          <Select
-            label={getLanguageByKey("typeMessages")}
-            placeholder={getLanguageByKey("typeMessages")}
-            data={MESSAGES_TYPE_OPTIONS}
-            value={mtype}
-            onChange={setMtype}
-            clearable
-          />
-
-          <DatePickerInput
-            type="range"
-            label={getLanguageByKey("searchByInterval")}
-            placeholder={getLanguageByKey("searchByInterval")}
-            value={timeSent}
-            onChange={setTimeSent}
-            valueFormat="DD-MM-YYYY"
-            clearable
-          />
-
-          <MultiSelect
-            label={getLanguageByKey("Selectează operator")}
-            placeholder={getLanguageByKey("Selectează operator")}
-            data={technicians}
-            value={senderIds}
-            onChange={setSenderIds}
-            searchable
-            clearable
-          />
-          <Flex justify="end" gap="md" mt="md">
-            <Button variant="outline" onClick={handleResetMessageFilter}>
-              {getLanguageByKey("Reset filter")}
-            </Button>
-            <Button variant="default" onClick={onClose}>
-              {getLanguageByKey("Închide")}
-            </Button>
-            <Button variant="filled" loading={loading} onClick={handleApplyMessageFilter}>
-              {getLanguageByKey("Aplică")}
-            </Button>
-          </Flex>
-          
-        </Flex>
+        <MessageFilterForm
+          initialData={initialData}
+          loading={loading}
+          onClose={onClose}
+          onSubmit={onSubmitTicket}
+        />
       </Tabs.Panel>
     </Tabs>
   );
